requirements.txt
Flask==3.0.0
requests==2.31.0
beautifulsoup4==4.12.2
Pillow==10.2.0
lxml==4.9.3
